changed the underlying webservice impl from wsdl2java to wsimport so it will not generate webservice impl on the fly which consumes alot of memory.
added more test cased and they all should be part of the regression test
from now on.

+++ synergyapps-parent/synergyapps-persistent/src/main/java/com/synel/synergy/synergy2416/persistent/EmployeeDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface EmployeeDao {
 	
-	public EmployeePOJO findEmployeeById(int id);
+	public EmployeePOJO findEmployeeById(long id);
 	
 	public EmployeePOJO findEmployeeByBadgeNumber(int BadgeNumber);
